Color palette: change hues in solarlike scheme
This is my second take at creating a color scheme based on solarized,
adjusted to my own preferences.  I think the result is quite good.

The basic idea behind solarized is excellent - I just don't like some
of the accent colors (in particular orange, green and violet). I have
adjusted them so that:
- orange is closer to yellow, and not so similar to red,
- green is much greener and not so yellowish,
- violet is significantly different than blue.

I have also changed the lightness of some of the accent colors relative
to their solarized versions:
- orange, yellow and cyan are brighter,
- blue is darker.

This may be questionable, as the idea behind solarized was to make
accent colors present an even contrast against the backgroud.  However,
I believe that we are so used to yellow and cyan being bright that it's
more difficult to recognize them when they are too dim.  I have checked
that they are still legible against the light version of the background.

Additionally
- red is just a bit more saturated,
- magenta is a tad less pink,
- cyan has slightly more green to distinguish from blue.

Finally, the foreground color is changed back to mach solarized scheme.