Color palette: introduce selenized color scheme
I've decided to give a name to the color scheme designed in the previous
commit, instead of calling it just "solarlike". To differentiate it from
solarized, I'm changing the background to a more blueish one (although it
should still have the same brightness).

I've also made the foreground color a bit brighter, but for now that's
specific to the application - I need slightly higher contrast for terminal.

Also, delete solarlike-violet as an unnecessary variation.