Keyboard: move right hand one column to the right
This is an attempt to gain one additional key accessible to my
thumbs, which are the strongest fingers.  I don't use right shift
and return keys for anything (because they're not very comfortable
to type); backslash key is easier to press, but I currently don't
use it, so I have little to loose.