Keyboard: swap D with V and move a lot of punctuation
I find QWERTY's E key surprisingly uncomfortable to type because of
how the key colums are staggered (i.e. E key is not exaclty above
D key, but slightly to the left).  I started developing pain in my
left hand because of this, so I'm switching D with less frequent V.