Remove my name from git configuration
I want this dotfiles repo to contain as little personal data as
possible.  In particular, I prefer to store my username in a separate
file deployed with ansible, so that if someone cloned this repository
(or used a VM on which I had installed these dotfiles myself) he won't
start making commits attributed to me.