Prompt: make the user command bold
This is done by leaving an "open" color code near the end of the prompt,
and then using the DEBUG trap (which is triggered when the user hits
Return) to close it.

I find that this dramatically improves readability and allows to parse
the history of a session at a glance.