Git logs: add --first-parent variants
These are very helpful for getting an overview of the situation in a
repository with many branches. If the team manages all changes in
feature branches, and merges them explicitly to the main branch, using
--first-parent will filter out history complexity while still giving
reader the essential information.

This is particularly useful for viewing multiple unmerged branches - for
example with `git lf --branches --remotes` you will see how the branches
depend on each other without seeing their internal history.