Git aliases: submodule support
By default "regular" git commands pretty much ignore any submodules that
are present in the repository, meaning that after every "git checkout"
etc. you have to run "git submodule update" (or maybe also init and
sync). Additionally, when a commit affects a submodule, an ordinary "git
diff" will just show you which hash was the HEAD of the submodule before
and after the commit. This is very inconvenient.

So, do the following:
- define `sup` alias for updating submodules (all possible steps)
- add a call to `git sup` after checkout, merge and rebase aliases
- add --submodule=log option to diff and log --patch aliases