Vim: try monitoring for external changes of the file
If an external appliation modifies the file that is edited, I want vim
to notice that and prompt me. Unfortunately, there are no good solutions
for this.

First, vim has a feature named "autoread", but it's very limited:
http://vi.stackexchange.com/questions/2702/how-can-i-make-vim-autoread-a-file-while-it-doesnt-have-focus

One can set additional triggers for the autoread check:
http://stackoverflow.com/a/18866818/2058424
but they're still not enough.

There's a plugin for this, but it has some issues:
https://github.com/Carpetsmoker/auto_autoread.vim/issues/3

Still, these changes should help.