Bash: use $HOME instead of ~ in PATH
This is a bit tricky. Using ~ in PATH will work, but only in shell,
because the shell expands ~ to $HOME (so, for example, my shell will
know about executable scripts from ~/bin). However, PATH is also used
directly by other programs, which don't do ~ expansion.

Note that bash doesn't expand ~ when exporting PATH because it's inside
quotes (which is necessary to prevent problems with e.g. spaces in dir
names).