Git: refactor 2-line log
Move decoration (i.e. branch names) to the beginning of the second line,
to lower the chance of them not fitting on the screen (also, they are
the second most iportant item displayed - right after commit subject -
so they should be given prominent place close to where the eye starts
looking for information).

Also, move committish to the first line - almost all logs have it like
this, and this also makes the second line (which in most cases contains
just author and date) visually weaker.