Git: autoguess email during clone
Get rid of cln, cln-janw and cln-clty aliases, as I haven't been using
them anyway. Instead, add "git-cln" script (git's subcommand name
resolution will make it available as "git cln", or "g cln" with my
alias) that tries to autodetect which email to use based on arguments
given to the clone command. This is not bulletproof, but it's intended
just as a helper, to avoid having to run "git gmail" in each and every
private project I will work on.