Git: remove redundant parts of color aliases
I used to add color reset after each element of the log to make sure that
coloring will affect only that single token; however, this is overly verbose
and not needed anymore. "Auto" color is really working well for everything
(except for the items I want to be in gray, which is mapped to "white" ANSI
code in my terminal).

I think "tformat" used to be necessary in the past to get multiline logs work
properly, but with new git versions it's no longer necessary so can be removed.