GPG: prompt for password in terminal
GPG agent is stupid and without this tries to spawn a graphical prompt
(even when I'm logged in via SSH), which errors and causes gpg itself to
fail with quite unhelpful

    gpg: public key decryption failed: Operation cancelled
    gpg: decryption failed: No secret key

error message.

I'd probably prefer to use terminal prompt only when necessary, but it
would be more work, so let's see whether this will be good enough.