Increase SSH key TTL in agent
Basically I want the key to be accesssible whole day. 10h meant that a
key added in the morning will expire before evening. On the other hand,
using 24h will be a bit confusing because key added on 10am would be
valid on 9am next day - better to always start the day with adding the
key.